Added enemies doing damage to townhall at basic level.
-also formatted townhall health as an int
-also added a gameover scnene when townhall reaches 0 and added it to build

+++ Assets/Scripts/UI Scripts/Townhall Info UI.cs
@@ -15,8 +15,8 @@ public class TownhallInfoUI : MonoBehaviour
 
     void Update()
     {
-        string townhallHealth = townhall.getHealth().ToString();
-        townhallHealthText.text = "Townhall Health: " + townhallHealth;
+        string townhallHealth = ((int) townhall.getHealth()).ToString();
+        townhallHealthText.text = "Townhall Health: " +  townhallHealth;
     }
 
 }
